Rilis 2412.1.0 (#8838)

+++ donjo-app/controllers/Admin_pembangunan.php
@@ -47,7 +47,8 @@ defined('BASEPATH') || exit('No direct script access allowed');
 
 class Admin_pembangunan extends Admin_Controller
 {
-    public $modul_ini = 'pembangunan';
+    public $modul_ini           = 'pembangunan';
+    public $kategori_pengaturan = 'Pembangunan';
 
     public function __construct()
     {
@@ -173,7 +174,7 @@ class Admin_pembangunan extends Admin_Controller
         redirect_with('error', 'Gagal Hapus Data');
     }
 
-    private function validasi($post, $id = null, $old_foto = null)
+    private function validasi(array $post, $id = null, ?string $old_foto = null): array
     {
         return [
             'sumber_dana'             => bersihkan_xss($post['sumber_dana']),
@@ -200,7 +201,7 @@ class Admin_pembangunan extends Admin_Controller
         ];
     }
 
-    private function upload_gambar_pembangunan(string $jenis, $old_foto = '')
+    private function upload_gambar_pembangunan(string $jenis, ?string $old_foto = null)
     {
         // Inisialisasi library 'upload'
         $this->load->library('MY_Upload', null, 'upload');
@@ -216,7 +217,7 @@ class Admin_pembangunan extends Admin_Controller
         $adaBerkas = ! empty($_FILES[$jenis]['name']);
         if (! $adaBerkas) {
             // Jika hapus (ceklis)
-            if (isset($_POST['hapus_foto'])) {
+            if (isset($_POST['hapus_foto']) && $old_foto !== null) {
                 unlink(LOKASI_GALERI . $old_foto);
 
                 return null;
